backend/pyproject.toml
[project]
name = "dreamfactory-backend"
version = "0.1.0"
description = "DreamFactory v2.1 Backend with Google ADK"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "google-genai>=0.2.0",
    "firebase-admin>=6.5.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.9.0",
    "pillow>=11.0.0",
    "requests>=2.32.0",
    # Assuming google-adk is available in the environment or a private index.
    # If it's not on PyPI, we might need a specific source or local install.
    # For now, listing it as a dependency.
    "google-adk>=0.1.0",
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.hatch.build.targets.wheel.force-include]
"." = "backend"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
